+++ test/watcher.js
@@ -4,7 +4,7 @@ const fs = require('fs-extra');
 const path = require('path');
 const {execSync} = require('child_process');
 
-let winfs
+let winfs;
 if (process.platform === 'win32') {
   winfs = require('@gyselroth/windows-fsstat');
 }
@@ -108,7 +108,7 @@ describe('watcher', () => {
 
       after(async () => {
         await sub.unsubscribe();
-        await fs.rmdir(tmpDir, { recursive: true });
+        await fs.rmdir(tmpDir, {recursive: true});
       });
 
       describe('files', () => {
@@ -147,10 +147,19 @@ describe('watcher', () => {
           await fs.rename(f1, f2);
 
           let res = await nextEvent();
-          assert.deepEqual(res, [
-            event({type: 'delete', path: f1, ino, fileId, kind}, {backend}),
-            event({type: 'create', path: f2, ino, fileId, kind}, {backend}),
-          ]);
+          if (backend === 'inotify' || backend === 'windows') {
+            assert.deepEqual(res, [
+              event(
+                {type: 'rename', oldPath: f1, path: f2, ino, fileId, kind},
+                {backend},
+              ),
+            ]);
+          } else {
+            assert.deepEqual(res, [
+              event({type: 'delete', path: f1, ino, fileId, kind}, {backend}),
+              event({type: 'create', path: f2, ino, fileId, kind}, {backend}),
+            ]);
+          }
         });
 
         it('should emit when an existing file is renamed', async () => {
@@ -159,21 +168,24 @@ describe('watcher', () => {
           await fs.rename(fileToRename, f2);
 
           let res = await nextEvent();
-          if (backend == 'windows') {
+          if (backend === 'inotify' || backend === 'windows') {
             assert.deepEqual(res, [
-              // The WindowsBackend does not have access to the removed file
-              // information and thus cannot tell us the delete event is for a
-              // file or give us its fileId.
               event(
-                {type: 'delete', path: fileToRename, kind: 'unknown'},
+                {
+                  type: 'rename',
+                  oldPath: fileToRename,
+                  path: f2,
+                  ino,
+                  fileId,
+                  kind,
+                },
                 {backend},
               ),
-              event({type: 'create', path: f2, fileId, kind}, {backend}),
             ]);
           } else {
             assert.deepEqual(res, [
               event({type: 'delete', path: fileToRename, ino, kind}, {backend}),
-              event({type: 'create', path: f2, ino, kind}, {backend}),
+              event({type: 'create', path: f2, ino, fileId, kind}, {backend}),
             ]);
           }
         });
@@ -188,10 +200,19 @@ describe('watcher', () => {
           await fs.rename(f1, f2);
 
           let res = await nextEvent();
-          assert.deepEqual(res, [
-            event({type: 'delete', path: f1, ino, fileId, kind}, {backend}),
-            event({type: 'create', path: f2, ino, fileId, kind}, {backend}),
-          ]);
+          if (backend === 'inotify' || backend === 'windows') {
+            assert.deepEqual(res, [
+              event(
+                {type: 'rename', oldPath: f1, path: f2, ino, fileId, kind},
+                {backend},
+              ),
+            ]);
+          } else {
+            assert.deepEqual(res, [
+              event({type: 'delete', path: f1, ino, fileId, kind}, {backend}),
+              event({type: 'create', path: f2, ino, fileId, kind}, {backend}),
+            ]);
+          }
         });
 
         it('should emit when a file is deleted', async () => {
@@ -231,10 +252,19 @@ describe('watcher', () => {
           await fs.rename(f1, f2);
 
           let res = await nextEvent();
-          assert.deepEqual(res, [
-            event({type: 'delete', path: f1, ino, fileId, kind}, {backend}),
-            event({type: 'create', path: f2, ino, fileId, kind}, {backend}),
-          ]);
+          if (backend === 'inotify' || backend === 'windows') {
+            assert.deepEqual(res, [
+              event(
+                {type: 'rename', oldPath: f1, path: f2, ino, fileId, kind},
+                {backend},
+              ),
+            ]);
+          } else {
+            assert.deepEqual(res, [
+              event({type: 'delete', path: f1, ino, fileId, kind}, {backend}),
+              event({type: 'create', path: f2, ino, fileId, kind}, {backend}),
+            ]);
+          }
         });
 
         it('should emit when an existing directory is renamed', async () => {
@@ -243,16 +273,19 @@ describe('watcher', () => {
           await fs.rename(dirToRename, f2);
 
           let res = await nextEvent();
-          if (backend === 'windows') {
-            // The WindowsBackend does not have access to the removed dir
-            // information and thus cannot tell us the delete event is for a
-            // directory or give us its fileId.
+          if (backend === 'inotify' || backend === 'windows') {
             assert.deepEqual(res, [
               event(
-                {type: 'delete', path: dirToRename, kind: 'unknown'},
+                {
+                  type: 'rename',
+                  oldPath: dirToRename,
+                  path: f2,
+                  ino,
+                  fileId,
+                  kind,
+                },
                 {backend},
               ),
-              event({type: 'create', path: f2, fileId, kind}, {backend}),
             ]);
           } else {
             assert.deepEqual(res, [
@@ -290,7 +323,7 @@ describe('watcher', () => {
             Math.random().toString(31).slice(2),
           );
           fs.mkdirpSync(dir);
-          let { ino, kind } = await getMetadata(dir);
+          let {ino, kind} = await getMetadata(dir);
           await new Promise((resolve) => setTimeout(resolve, 100));
 
           let sub = await watcher.subscribe(dir, fn, {backend});
@@ -367,10 +400,19 @@ describe('watcher', () => {
           await fs.rename(f2, f3);
 
           let res = await nextEvent();
-          assert.deepEqual(res, [
-            event({type: 'delete', path: f2, ino, fileId, kind}, {backend}),
-            event({type: 'create', path: f3, ino, fileId, kind}, {backend}),
-          ]);
+          if (backend === 'inotify' || backend === 'windows') {
+            assert.deepEqual(res, [
+              event(
+                {type: 'rename', oldPath: f2, path: f3, ino, fileId, kind},
+                {backend},
+              ),
+            ]);
+          } else {
+            assert.deepEqual(res, [
+              event({type: 'delete', path: f2, ino, fileId, kind}, {backend}),
+              event({type: 'create', path: f3, ino, fileId, kind}, {backend}),
+            ]);
+          }
         });
 
         it('should emit when a sub-file is deleted', async () => {
@@ -420,10 +462,19 @@ describe('watcher', () => {
           await fs.rename(f2, f3);
 
           let res = await nextEvent();
-          assert.deepEqual(res, [
-            event({type: 'delete', path: f2, ino, fileId, kind}, {backend}),
-            event({type: 'create', path: f3, ino, fileId, kind}, {backend}),
-          ]);
+          if (backend === 'inotify' || backend === 'windows') {
+            assert.deepEqual(res, [
+              event(
+                {type: 'rename', oldPath: f2, path: f3, ino, fileId, kind},
+                {backend},
+              ),
+            ]);
+          } else {
+            assert.deepEqual(res, [
+              event({type: 'delete', path: f2, ino, fileId, kind}, {backend}),
+              event({type: 'create', path: f3, ino, fileId, kind}, {backend}),
+            ]);
+          }
         });
 
         it('should emit when a sub-directory is deleted with files inside', async () => {
@@ -575,17 +626,8 @@ describe('watcher', () => {
             assert.deepEqual(res, [
               event(
                 {
-                  type: 'delete',
-                  path: getPath('dir'),
-                  ino: dirIno,
-                  fileId: dirFileId,
-                  kind: dirKind,
-                },
-                {backend},
-              ),
-              event(
-                {
-                  type: 'create',
+                  type: 'rename',
+                  oldPath: getPath('dir'),
                   path: getPath('dir2'),
                   ino: dirIno,
                   fileId: dirFileId,
@@ -597,12 +639,34 @@ describe('watcher', () => {
               // information and thus cannot give us its inode but it can tell
               // us this is an event for a directory.
               event(
-                {type: 'delete', path: getPath('dir2/subdir'), kind: subdirKind},
+                {
+                  type: 'rename',
+                  oldPath: getPath('dir2/subdir'),
+                  path: getPath('dir2/subdir2'),
+                  ino: subdirIno,
+                  fileId: subdirFileId,
+                  kind: subdirKind,
+                },
+                {backend},
+              ),
+            ]);
+          } else if (backend === 'inotify' || backend === 'windows') {
+            assert.deepEqual(res, [
+              event(
+                {
+                  type: 'rename',
+                  oldPath: getPath('dir'),
+                  path: getPath('dir2'),
+                  ino: dirIno,
+                  fileId: dirFileId,
+                  kind: dirKind,
+                },
                 {backend},
               ),
               event(
                 {
-                  type: 'create',
+                  type: 'rename',
+                  oldPath: getPath('dir2/subdir'),
                   path: getPath('dir2/subdir2'),
                   ino: subdirIno,
                   fileId: subdirFileId,
@@ -698,17 +762,8 @@ describe('watcher', () => {
           assert.deepEqual(res, [
             event(
               {
-                type: 'delete',
-                path: getPath('dir'),
-                ino: dirIno,
-                fileId: dirFileId,
-                kind: dirKind,
-              },
-              {backend},
-            ),
-            event(
-              {
-                type: 'create',
+                type: 'rename',
+                oldPath: getPath('dir'),
                 path: getPath('dir2'),
                 ino: dirIno,
                 fileId: dirFileId,
@@ -777,10 +832,19 @@ describe('watcher', () => {
           await fs.rename(f2, f3);
 
           let res = await nextEvent();
-          assert.deepEqual(res, [
-            event({type: 'delete', path: f2, ino, fileId, kind}, {backend}),
-            event({type: 'create', path: f3, ino, fileId, kind}, {backend}),
-          ]);
+          if (backend === 'inotify' || backend === 'windows') {
+            assert.deepEqual(res, [
+              event(
+                {type: 'rename', oldPath: f2, path: f3, ino, fileId, kind},
+                {backend},
+              ),
+            ]);
+          } else {
+            assert.deepEqual(res, [
+              event({type: 'delete', path: f2, ino, fileId, kind}, {backend}),
+              event({type: 'create', path: f3, ino, fileId, kind}, {backend}),
+            ]);
+          }
         });
 
         it('should emit when a symlink is deleted', async () => {
@@ -898,14 +962,31 @@ describe('watcher', () => {
             let f4 = getFilename();
             await fs.writeFile(f1, 'hello world');
             let {ino, fileId, kind} = await getMetadata(f1);
+            await nextEvent();
+
             await fs.rename(f1, f2);
             await fs.rename(f2, f3);
             await fs.rename(f3, f4);
 
             let res = await nextEvent();
-            assert.deepEqual(res, [
-              event({type: 'create', path: f4, ino, fileId, kind}, {backend}),
-            ]);
+            if (backend === 'inotify' || backend === 'windows') {
+              assert.deepEqual(res, [
+                event(
+                  {type: 'rename', oldPath: f1, path: f4, ino, fileId, kind},
+                  {backend},
+                ),
+              ]);
+            } else if (backend === 'watchman') {
+              assert.deepEqual(res, [
+                event({type: 'create', path: f4, ino, fileId, kind}, {backend}),
+                event({type: 'delete', path: f1, ino, fileId, kind}, {backend}),
+              ]);
+            } else {
+              assert.deepEqual(res, [
+                event({type: 'delete', path: f1, ino, fileId, kind}, {backend}),
+                event({type: 'create', path: f4, ino, fileId, kind}, {backend}),
+              ]);
+            }
           });
         }
